refactor do dbAction in HttpWsFitness.kt fix failing tests
