+++ core/src/main/kotlin/org/evomaster/core/problem/rest/service/RestFitness.kt
@@ -36,12 +36,13 @@ open class RestFitness : AbstractRestFitness<RestIndividual>() {
                 individual.seeActions().size)
         }
 
-        doInitializingActions(individual)
+        val actionResults: MutableList<ActionResult> = mutableListOf()
+
+        doDbCalls(individual.seeInitializingActions(), actionResults = actionResults)
 
 
         val fv = FitnessValue(individual.size().toDouble())
 
-        val actionResults: MutableList<ActionResult> = mutableListOf()
 
         //used for things like chaining "location" paths
         val chainState = mutableMapOf<String, String>()
