+++ core/src/main/kotlin/org/evomaster/core/problem/rest/service/RestResourceFitness.kt
@@ -126,89 +126,7 @@ class RestResourceFitness : AbstractRestFitness<RestIndividual>() {
 
     }
 
-    /**
-     * @param allDbActions specified the db actions to be executed
-     * @param sqlIdMap indicates the map id of pk to generated id
-     * @param allSuccessBefore indicates whether all SQL before this [allDbActions] are executed successfully
-     * @param previous specified the previous db actions which have been executed
-     * @return whether [allDbActions] execute successfully
-     */
-    private fun doDbCalls(allDbActions : List<DbAction>,
-                          sqlIdMap : MutableMap<Long, Long>,
-                          allSuccessBefore : Boolean,
-                          previous: MutableList<DbAction>,
-                          actionResults: MutableList<ActionResult>
-    ) : Boolean {
 
-        if (allDbActions.isEmpty()) {
-            return true
-        }
-
-
-        val dbresults = (allDbActions.indices).map { DbActionResult() }
-
-        if (allDbActions.none { !it.representExistingData }) {
-            /*
-                We are going to do an initialization of database only if there
-                is data to add.
-                Note that current data structure also keeps info on already
-                existing data (which of course should not be re-inserted...)
-             */
-            // other dbactions might bind with the representExistingData, so we still need to record sqlId here.
-            allDbActions.filter { it.representExistingData }.flatMap { it.seeGenes() }.filterIsInstance<SqlPrimaryKeyGene>().forEach {
-                sqlIdMap.putIfAbsent(it.uniqueId, it.uniqueId)
-            }
-            previous.addAll(allDbActions)
-            actionResults.addAll(dbresults)
-            return true
-        }
-
-        val dto = try {
-            DbActionTransformer.transform(allDbActions, sqlIdMap, previous)
-        }catch (e : IllegalArgumentException){
-            // the failure might be due to previous failure
-            if (!allSuccessBefore){
-                previous.addAll(allDbActions)
-                return false
-            } else
-                throw e
-        }
-        dto.idCounter = StaticCounter.getAndIncrease()
-
-        val sqlResults = rc.executeDatabaseInsertionsAndGetIdMapping(dto)
-        val map = sqlResults?.idMapping
-        val executedResults = sqlResults?.executionResults
-
-        if (executedResults?.size?:0 > allDbActions.size)
-            throw IllegalStateException("incorrect insertion execution results (${executedResults!!.size}) which is more than the size of insertions (${allDbActions.size}).")
-        executedResults?.forEachIndexed { index, b ->
-//            allDbActions[index].insertExecutedSuccessfully = b
-            dbresults[index].setInsertExecutionResult(b)
-        }
-        actionResults.addAll(dbresults)
-        previous.addAll(allDbActions)
-
-
-        if (map == null) {
-            LoggingUtil.uniqueWarn(log, "Failed in executing database command")
-            return false
-        }else{
-            val expected = allDbActions.filter { !it.representExistingData }
-                .flatMap { it.seeGenes() }.flatMap { it.flatView() }
-                .filterIsInstance<SqlPrimaryKeyGene>()
-                .filter { it.gene is SqlAutoIncrementGene }
-                .filterNot { it.gene is SqlForeignKeyGene }
-            val missing = expected.filterNot {
-                map.containsKey(it.uniqueId)
-            }
-            sqlIdMap.putAll(map)
-            if (missing.isNotEmpty()){
-                log.warn("can not get sql ids for {} from sut", missing.map { "${it.uniqueId} of ${it.tableName}" }.toSet().joinToString(","))
-                return false
-            }
-        }
-        return true
-    }
 
     override fun hasParameterChild(a: RestCallAction): Boolean {
         return sampler.seeAvailableActions()
